requirements.txt
accelerate
pycls
diffusers
einops
ninja
pycocotools==2.0.7
numba
numpy
opencv-python
pandas==1.4.4
pillow==10.2.0
protobuf
scikit-image
scikit-learn
scipy
seaborn
Shapely
six
tensorboard
timm
tqdm
lyft_dataset_sdk
nuscenes-devkit
casadi
cython
addict
packaging
matplotlib
prettytable
terminaltables
cityscapesscripts
imagecorruptions
open3d
networkx
ipython
# metric related
motmetrics==1.1.3 # Fixed
trimesh==2.35.39
# pytest related
pytest
pytest-cov
pytest-runner
yapf
flake8
similaritymeasures
laspy
# lazrs
py-trees
simple_watchdog_timer
transforms3d
tabulate
ephem
dictor
torchdiffeq